chore: consume the extracted github.com/richardwooding/c2pa module
internal/content/c2pa.go has been extracted to its own pure-Go library
(github.com/richardwooding/c2pa v0.1.0 — read-only C2PA / Content
Credentials reader: JUMBF + CBOR claim + COSE_Sign1 signer + RFC 3161
signing time). Depend on the published module instead of a private copy:
single source of truth, and it dogfoods the library we ship to the
community.

- require github.com/richardwooding/c2pa v0.1.0; fxamacker/cbor +
  veraison/go-cose drop to indirect (only the module uses them directly).
- repoint the one consumer (imagetype.go): extractC2PA(ctx, container, rs)
  -> c2pa.Read(ctx, container, rs); c2paContainer now returns
  (c2pa.Container, bool). The is_c2pa / c2pa_* attributes are byte-for-byte
  identical.
- delete internal/content/c2pa.go + its pure-parser tests + the 3 fuzz
  targets (they moved to the module, which carries its own fuzz CI).
- keep the integration test (now c2pa_integration_test.go) — it exercises
  imageType.Attributes, not the parser — and the c2pa_signed.jpg fixture.
- drop the 3 C2PA entries from fuzz.yml; update CLAUDE.md.

C2PA behaviour is unchanged — full suite passes under -race; vet, go fix,
lint clean; --list and a live search still surface c2pa_signed_by /
c2pa_signed_at against the fixture.
