+++ CLAUDE.md
@@ -44,7 +44,7 @@ Empty `Expr` defaults to `"true"` (matches all files). Worker count defaults to 
 
 ### Fuzz testing
 
-High-risk parsers (frontmatter, MP3 ID3v2 + Xing, MKV EBML, MP4 box walker, C2PA JUMBF box walker + COSE_Sign1 + RFC 3161 ASN.1 timestamp, office/EPUB/email/PDF body extractors, CEL compile, index gob decoder, every hand-rolled binary header walker) have `FuzzXxx` targets in `*_fuzz_test.go`.
+High-risk parsers (frontmatter, MP3 ID3v2 + Xing, MKV EBML, MP4 box walker, office/EPUB/email/PDF body extractors, CEL compile, index gob decoder, every hand-rolled binary header walker) have `FuzzXxx` targets in `*_fuzz_test.go`. (The C2PA JUMBF/COSE/RFC 3161 parser was extracted to `github.com/richardwooding/c2pa`, which carries its own fuzz targets.)
 
 - **Regular CI** (`ci.yml`): `go test ./...` runs each target's seed corpus — a seed panic fails the build.
 - **Scheduled fuzz** (`fuzz.yml`): 5 min/target nightly at 03:30 UTC. Crashes land in `testdata/fuzz/<FuzzName>/<hash>` — commit those to lock in regression coverage. Manual `workflow_dispatch` accepts a `fuzztime` input for bug-hunt sessions.
@@ -64,7 +64,7 @@ Six internal packages compose the pipeline. The first three are tightly coupled 
 
 **Robustness invariants (issue #337, pinned by tests — see `internal/search/doc.go`):** (1) *Graceful degradation* — a detected file whose `Attributes()` errors/panics still appears in `search 'true'` with its base attributes; never silently dropped. (2) *Cancellation* — every orchestrator and every hand-rolled parser loop checks `ctx.Err()` at entry and inside any unbounded loop, returning the ctx error (streaming) or `Cancelled=true` (buffered aggregators) promptly. `TestWalk_GracefulDegradation_EveryType` and `TestOrchestrators_PreCancelledCtxReturnPromptly` fail the suite if either regresses — uphold both when adding a content type or an orchestrator/loop.
 
-- **`internal/content`** — pluggable content-type detection. Each type implements `ContentType` (`Name`, `Extensions`, `MagicBytes`, `Attributes(ctx, path)`) and self-registers in `init()` via `Register(...)`. `Registry.Detect(path)` tries exact-basename match (optional `FilenameMatcher` interface), then longest-suffix extension match, then magic-byte sniffing on the first 512 bytes. `Attributes` is called per matching file and returns `map[string]any`; implementations check `ctx.Err()` at entry and inside any unbounded loop. Covers: markdown, json/xml/html/yaml/toml, pdf, image variants (EXIF via `imagemeta`), text/csv, epub, office (DOCX/XLSX/PPTX/ODT), audio (tags via `dhowden/tag`, playback via hand-rolled per-format), video (MP4/MKV/AVI), archive (ZIP/TAR/TAR.GZ/GZIP), binary (ELF/Mach-O/PE via `debug/*`), email (RFC 5322 + mbox), source (19 languages), notebooks (Jupyter/Zeppelin), disk images (DMG/ISO/VHD/VHDX/VMDK/QCOW2/WIM), install packages (PKG/DEB/RPM/AppImage), VM bytecode (Java class/Python pyc/WASM), science data (FITS/VOTable/HDF5/PDS3+4/CDF), database (SQLite), font (SFNT family), and repo/build/system metadata files. Bodies are extracted on demand by `internal/content/body.go` (`ExtractBody`) — office walks ZIP+XML, email walks MIME, PDF uses `ledongthuc/pdf` wrapped in defer/recover. `setTypeFlags` in `internal/celexpr/typeflags.go` is the single source of truth for content-type-name → `is_X` predicate dispatch.
+- **`internal/content`** — pluggable content-type detection. Each type implements `ContentType` (`Name`, `Extensions`, `MagicBytes`, `Attributes(ctx, path)`) and self-registers in `init()` via `Register(...)`. `Registry.Detect(path)` tries exact-basename match (optional `FilenameMatcher` interface), then longest-suffix extension match, then magic-byte sniffing on the first 512 bytes. `Attributes` is called per matching file and returns `map[string]any`; implementations check `ctx.Err()` at entry and inside any unbounded loop. Covers: markdown, json/xml/html/yaml/toml, pdf, image variants (EXIF via `imagemeta`), text/csv, epub, office (DOCX/XLSX/PPTX/ODT), audio (tags via `dhowden/tag`, playback via hand-rolled per-format), video (MP4/MKV/AVI), archive (ZIP/TAR/TAR.GZ/GZIP), binary (ELF/Mach-O/PE via `debug/*`), email (RFC 5322 + mbox), source (19 languages), notebooks (Jupyter/Zeppelin), disk images (DMG/ISO/VHD/VHDX/VMDK/QCOW2/WIM), install packages (PKG/DEB/RPM/AppImage), VM bytecode (Java class/Python pyc/WASM), science data (FITS/VOTable/HDF5/PDS3+4/CDF), database (SQLite), font (SFNT family), and repo/build/system metadata files. Bodies are extracted on demand by `internal/content/body.go` (`ExtractBody`) — office walks ZIP+XML, email walks MIME, PDF uses `ledongthuc/pdf` wrapped in defer/recover. C2PA / Content Credentials (the `is_c2pa` / `c2pa_*` image attributes) are read by the extracted `github.com/richardwooding/c2pa` module (`c2pa.Read` in `imagetype.go`) — pure-Go, read-only, unverified. `setTypeFlags` in `internal/celexpr/typeflags.go` is the single source of truth for content-type-name → `is_X` predicate dispatch.
 
 - **`internal/celexpr`** — wraps `cel-go`. `New(expr)` declares a fixed schema of CEL variables (union of common file attrs + every type-specific attribute) and compiles the program once. `BuildAttributesWith(ctx, fsys, fsPath, displayPath, registry, opts)` is the primary attribute-extraction entry point: runs `os.Stat`, optionally consults `opts.Index` (a `(size, mtime)`-validated cache), calls `registry.Detect` + `ContentType.Attributes` only on cache miss, and packs into `FileAttributes`. `Evaluate` flattens `FileAttributes` into a CEL activation, supplying zero values for type-specific vars (cel-go errors on undeclared/unbound). `Schema()` (in `schema.go`) returns the structured docs that drive both `--list` and the MCP `list_attributes` tool — the only place attribute docs are hard-coded. Symlink awareness lives in `probeSymlink` / `applySymlinkInfo`.
